ASoC: CSR: atlas7b: persistent storage for kas
Persistent Storage (PS) from the DSP point of view is way for
the DSP to restore settings after a power cycle or system reset.
The DSP Persistent Storage is represented as a binary file located
in the ARM's filesystem and is loaded into RAM at a boot up time.
When the DSP is loaded with its firmware code and started, the memory
area where the PS file is loaded has to be accessible by the DSP
subsystem. The physical address in RAM where the PS file is mapped
is also stored to a DSP's Data Memory location to make aware the DSP
of the PS RAM region. During the run time the DSP may issue at some
point a FLUSH_REQ to demand the data from the PS memory to be flushed
back to the disk. Usually, once data structures in PS RAM are updated
by the DSP, the DSP will issue a FLUSH_REQ. ARM system will respond
back to the DSP with a FLUSH_RES to indicate the completion of
the sych process to the disk.

Signed-off-by: Andreas Orfanos <aorfan@codeaurora.org>
Change-Id: I79e2cad599d42fa27a42041066ed96c16d62704a