mmc: sirf: modify register read function
The register contents related with host version and capability
in SD_SLOT_INT_STATUS(0xFC) and SDHCI_CAPABILITIES_1(0x44) change to
the values represting the real capability of our host controller. So
we don't need to hack them anymore. While there is still a mistake,
SDHCI_USE_SDR50_TUNING bit is forgotten and this will affect the
performance of SDR50 card.

Change-Id: I2e30eaca72b0ec44464beb9cdb4f22ec0c00ad40
Signed-off-by: Weijun Yang <weijuny@codeaurora.org>