ASoC: CSR: atlas7b: Load the constant data into DRAM, Kalimba will use it.
Some of the operators need big chunks of data to be stored somewhere,
things like filter coefficients. If we take resampler as the easiest
example there are different combinations of upsample/downsample and
different quality, all this combination use different table of filter
coefficients.

Before this were stored in our Data memory (SRAM) using a lot of it
although in most of the cases we are not using them, so storing these
filters on DRAM allows KAS to have much more data memory that can be
used for processing audio data.
When a filter is needed it is copied from DRAM to SRAM and used,
and as soon as it is not needed anymore it is freed from KAS SRAM.

Change-Id: I1bed45eeca072f0a6786ea856156197259b43d6f
Signed-off-by: Rongjun Ying <rying@codeaurora.org>