ARM:dts: atlas7: Change the mux group of SD2 WP
We've changed the mux group from 0 to 1. While there came an issue,
the default value of gpio_4 filed of pinctl is 0x2, which points to
SD2 WP. If we use mux group 1 by setting in dts, the jtag_trstn filed
is 0x7, which also points to SD2 WP. The result is SD2 WP is connected
to two pad simultaneously. This may lead the WP status to unexpectable
state.
So, we step back to use mux0.

Change-Id: If7e76306efcd1cb9a0838e0bf21d8990d3c6153e
Signed-off-by: Weijun Yang <weijuny@codeaurora.org>