ASoC: Atlas7: move kcoredump to a separate thread context
With the current implementation of the coredump, if the
application dies, becomes zombie, the coredump may not
happen because related structures connected with the
running application may have been freed. The solution to
this problem is to kick off the coredump as an action to the
system delay work queue thread, and let it be handled there.

Signed-off-by: Andreas Orfanos <aorfan@codeaurora.org>
Change-Id: I97cf7d8c786bb6780e41e5a0434304e4f61034b6