ASoC: Atlas7: kas coredump header correction
It has been suggested to make changes to kas coredump header.
The new header has to have the following format:

XCD2
AV 00600000
P DSP
AT KALIMBA5

AV directive is moved to the second from the top, and the most
significant 16bit corresponds to Global Version Register,
always 0x0060. And the less significant 16bits corresponds
to the chip version ID. That is read as 0x19.
The new hardcoded version in the coredump header should be
0x00600019

Signed-off-by: Andreas Orfanos <aorfan@codeaurora.org>
Change-Id: I075be8a3ad5499cfecff5d42ddda647574dee707