ASoC: CSR: KAS: implement mixer
What makes mixer tricky is it has a so called primary stream, which must
be set correctly to work. Say, you start one stream, and set the primary
stream to 1. After a while a second stream starts. If the first stream
stops earlier than the second one, you must correct primary stream to 2.
Combining all three inputs and random start/stop sequences, it will be a
nightmare to maintain.

The solution is to save operator's endpoint status and tell the operator
when the connection status changes. So the operator always knows the
latest activity of its endpoints and can take actions by itself. For
mixer, it can use this information to easily determine if primary stream
should be updated.

Last question: Why can't Kalimba firmware take care of this mess?

Change-Id: I4885f4cb29199772024745ffac272702938ea9e1
Signed-off-by: Yibo Cai <yiboc@codeaurora.org>