ASoC: CSR: KAS: support alarm stream
Just as expected, no single line of coding is required to support
this new stream.

To manage audio pipelines manually is possible, but it requires
documenting the diagram and namings very carefully, and it's still
an error prone job. For customer using, a GUI based configuration
tool is absolutely required.

Change-Id: Iee5c3ddc05504adeb771eab8742049ff2028eb14
Signed-off-by: Yibo Cai <yiboc@codeaurora.org>