media: vout: fix color control problem.
This commit is to fix bug:
APBU-11545 [A7DA][VPP]ColorAdjust effect is wrong after set
negative value then set right.
The resolution is when opending video device,
v4l2_ctrl_handler_setup should be called to ensure that both the
internal data structures and the hardware are in sync.

Change-Id: Ic9f3aeb2a45fddf414501f36202888a57fa25355
Signed-off-by: Guchun Chen <guchunc@codeaurora.org>