pwm: sirf: enable pwm first when resuming
Some pwm consumer do not implement suspend/resume function in
its driver, so pwm driver will do this for it.

Here is a corner case,
1. system suspends.
2. some pwm consumer does nothing.
2. pwm driver suspends, state of some pwm channel is "ENABLED".
3. system resumes.
4. some pwm consumer does nothing.
5. pwm driver resumes, some pwm channel will not be enabled indeed
because of state of this channel is already "ENABLED", the clock of
pwm controller is also not enabled, then write to register of pwm
will lead to hang.

So, we should enable pwm first when resuming.

Change-Id: Ia48a221cb9f6c044e7c5ca0d39db154d13ef1a14
Signed-off-by: Huayi Li <huayil@codeaurora.org>