pm:sirf: clear IPC register if M3 in HOLD state
The M3 could be in HOLD state in independent mode
for recovery with the HOLD flag in IPC register.
In the HOLD state, M3 never response the common
IPC. The Atlas7 clear the HOLD flag as a special
IPC to request M3 do reset in that special state.
This new feature is implemented the auto reset after
recovery finished in independent mode.

Change-Id: Icc927b0503bfa921a97e3b0adee1bc25749a9cfa
Signed-off-by: Ye He <yeh@codeaurora.org>