iio:adc:Add mutex to prevent code reenter
All the ADC channels, including AUX measurement, temperature sensor
and touch screen share one control register. While there is no
protection to the register accessing. So we may encounter the issue
that when we try to read back the value of one AUX channel, but if
another channel's read request is triggered, then the content of
request will be changed immediately. So the wrong value will be
returned.
So we add a mutex to make sure the read back process is atomic.

Change-Id: I9932ec9f39de15fc62e5da097c743668cfaab66c
Signed-off-by: Weijun Yang <weijuny@codeaurora.org>