ASoC: CSR: KAS: Adjust the sequence of pipeline stop
Links stop should be first to be stop, which will handle the primary stream
changing. If not, it will cause the mixer to consume it’s data until it is
empty, then changing the primary won’t save this situation, anyway what
should happen is changing the primary before stopping the other stream.

Change-Id: I83ff01be258d117209106e1389cfda5c54644754
Signed-off-by: Tony Han <xiahan@codeaurora.org>