ADC:sirf:Verify various sgain logic
For now only the voltage calculate algorithms of s_gain 0 and 2
are verified. The ones for s_gain 1,3,4,5,6,7 are still unfinished.
While these gain seetings may be useful for the final user.
With the gain error value stored in OTP table, we are able to
complete the algorithms. All the gain settings of the channels
are tested.

Change-Id: I9913ba81556c8793ad823cd9d91f1c4b49d664d4
Signed-off-by: Weijun Yang <weijuny@codeaurora.org>